requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==0.4.6
altair==4.2.0
antlr4-python3-runtime==4.8
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
blinker==1.5
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
-e git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1#egg=clip
coloredlogs==15.0.1
commonmark==0.9.1
contourpy==1.0.4
cycler==0.11.0
decorator==5.1.1
diffusers==0.2.4
einops==0.3.0
entrypoints==0.4
filelock==3.8.0
Flask==2.2.2
flatbuffers==2.0.7
fonttools==4.37.1
frozenlist==1.3.1
fsspec==2022.8.2
ftfy==6.1.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.49.0rc1
huggingface-hub==0.9.1
humanfriendly==10.0
idna==3.3
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.4.0
importlib-metadata==4.12.0
invisible-watermark==0.1.5
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.15.0
kiwisolver==1.4.4
kornia==0.6.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0rc2
mpmath==1.2.1
multidict==6.0.2
networkx==2.8.6
numpy==1.23.1
oauthlib==3.2.0
omegaconf==2.1.1
onnx==1.12.0
onnxruntime==1.12.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.5.0rc0
Pillow==9.2.0
protobuf==3.19.4
pudb==2019.2
pyarrow==9.0.0
pyasn1==0.5.0rc2
pyasn1-modules==0.3.0rc1
pycodestyle==2.9.1
pydeck==0.8.0b1
pyDeprecate==0.3.1
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-datauri==1.1.0
python-dateutil==2.8.2
pytorch-lightning==1.4.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
scikit-image==0.19.3
scipy==1.9.1
semver==3.0.0.dev3
shapely==2.0a1
six==1.16.0
smmap==5.0.0
streamlit==1.12.2
sympy==1.11.1
-e git+https://github.com/CompVis/taming-transformers.git@24268930bf1dce879235a7fddd0b2355b84d7ea6#egg=taming_transformers
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
test-tube==0.7.5
tifffile==2022.8.12
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.0
torch==1.13.0.dev20220901
torch-fidelity==0.3.0
torchmetrics==0.6.0
torchvision==0.14.0.dev20220901
tornado==6.2
tqdm==4.64.0
transformers==4.19.2
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
urwid==2.1.2
validators==0.20.0
wcwidth==0.2.5
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.8.1
